inline `optimizeCss` in leaf packages
Instead of introducing a custom `@tailwindcss/optimize` package
