--- packages/@tailwindcss-optimize/package.json
@@ -1,39 +1,0 @@
-{
-  "name": "@tailwindcss/optimize",
-  "version": "4.0.0-alpha.8",
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/tailwindlabs/tailwindcss.git",
-    "directory": "packages/@tailwindcss-optimize"
-  },
-  "bugs": "https://github.com/tailwindlabs/tailwindcss/issues",
-  "homepage": "https://tailwindcss.com",
-  "scripts": {
-    "lint": "tsc --noEmit",
-    "build": "tsup-node ./src/index.ts --format esm,cjs --dts --cjsInterop --splitting --minify --clean",
-    "dev": "pnpm run build -- --watch"
-  },
-  "files": [
-    "dist/"
-  ],
-  "exports": {
-    ".": {
-      "types": "./src/index.ts",
-      "require": "./dist/index.js",
-      "import": "./src/index.ts"
-    }
-  },
-  "publishConfig": {
-    "provenance": true,
-    "access": "public",
-    ".": {
-      "types": "./dist/index.d.ts",
-      "require": "./dist/index.js",
-      "import": "./dist/index.mjs"
-    }
-  },
-  "dependencies": {
-    "lightningcss": "^1.24.0"
-  }
-}
