+++ packages/@tailwindcss-vite/src/index.ts
@@ -1,5 +1,5 @@
-import { optimizeCss } from '@tailwindcss/optimize'
 import { IO, Parsing, scanFiles } from '@tailwindcss/oxide'
+import { Features, transform } from 'lightningcss'
 import path from 'path'
 import { compile } from 'tailwindcss'
 import type { Plugin, Update, ViteDevServer } from 'vite'
@@ -181,4 +181,28 @@ export default function tailwindcss(): Plugin[] {
       },
     },
   ] satisfies Plugin[]
+}
+
+function optimizeCss(
+  input: string,
+  { file = 'input.css', minify = false }: { file?: string; minify?: boolean } = {},
+) {
+  return transform({
+    filename: file,
+    code: Buffer.from(input),
+    minify,
+    sourceMap: false,
+    drafts: {
+      customMedia: true,
+    },
+    nonStandard: {
+      deepSelectorCombinator: true,
+    },
+    include: Features.Nesting,
+    exclude: Features.LogicalProperties,
+    targets: {
+      safari: (16 << 16) | (4 << 8),
+    },
+    errorRecovery: true,
+  }).code.toString()
 }
