+++ packages/tailwindcss/tests/ui.spec.ts
@@ -1,9 +1,9 @@
 import { expect, test, type Page } from '@playwright/test'
-import { optimizeCss } from '@tailwindcss/optimize'
 import { IO, Parsing, scanFiles } from '@tailwindcss/oxide'
 import fs from 'fs'
 import path from 'path'
 import { compile } from '../src'
+import { optimizeCss } from '../src/test-utils/run'
 
 const html = String.raw
 const css = String.raw
